bug-hunt #567: btrfs - drop 32-bit LBA truncation in file read path on large disks
btrfs_read extent-data loop and btrfs_read_blocks cast the computed
uint64_t LBA to uint32_t when calling blockdev_read_sectors, silently
truncating sector addresses on volumes over 2^32 sectors (>2TB).
Extent data mapped to a physical offset beyond 2TB would read the
wrong sector. Matches the LBA-truncation fix already applied to
btrfs_read_node/csum-verify (#800) and ext4 (#565).

+++ src/fs/btrfs.c
@@ -48,7 +48,7 @@ static int btrfs_read_blocks(struct btrfs_priv *bp, uint64_t bytenr,
     uint32_t sectors = count * 512 / 512; /* count is in 512-byte sectors */
     /* but count may be in nodesize/sectorsize units */
     for (uint32_t i = 0; i < count; i++) {
-        if (blockdev_read_sectors(bp->dev_id, (uint32_t)(lba + i), 1,
+        if (blockdev_read_sectors(bp->dev_id, lba + i, 1,
                                    buf + i * 512) != 0)
             return -1;
     }
@@ -1743,7 +1743,7 @@ static int btrfs_read(void *priv, const char *path,
                     uint8_t *dest = (uint8_t *)buf + done;
 
                     for (uint32_t s = 0; s < nsect; s++) {
-                        if (blockdev_read_sectors(bp->dev_id, (uint32_t)(lba + s), 1,
+                        if (blockdev_read_sectors(bp->dev_id, lba + s, 1,
                                                    dest + s * 512) != 0) {
                             /* I/O error — return what we have */
                             goto read_done;
